FastAPI/requirements.txt
fastapi
uvicorn
pydantic
mlflow
torch
torchvision
Pillow
boto3
python-multipart